refactor: consolidate admin resources into dedicated admin module
Move admin-organizations/ and api-keys/ into src/modules/admin/ with a
composite controller (prefix /v1/admin). Sub-controllers now use relative
prefixes, inheriting from the parent. Routes remain unchanged.

Closes #108
