+++ src/test/helpers/app.ts
@@ -5,8 +5,7 @@ import { betterAuthPlugin } from "@/lib/auth-plugin";
 import { errorPlugin } from "@/lib/errors/error-plugin";
 import { healthPlugin } from "@/lib/health";
 import { loggerPlugin } from "@/lib/logger";
-import { adminOrganizationsController } from "@/modules/admin-organizations";
-import { apiKeysController } from "@/modules/api-keys";
+import { adminController } from "@/modules/admin";
 import { auditController } from "@/modules/audit";
 import { employeeController } from "@/modules/employees";
 import { occurrencesController } from "@/modules/occurrences";
@@ -27,13 +26,12 @@ export function createTestApp() {
       })
     )
     .use(betterAuthPlugin)
-    .use(adminOrganizationsController)
+    .use(adminController)
     .use(organizationController)
     .use(employeeController)
     .use(occurrencesController)
     .use(paymentsController)
     .use(auditController)
-    .use(apiKeysController)
     .get("/", ({ redirect }) => redirect("/health"));
 }
